Calculate shuffle sharding ingester lookback based on override
The shuffle sharding lookback value is calculated based on whether
shuffle sharding is enabled and whether query-ingesters-within is
greater than zero. As we've changed to using a tenant override for
getting query-ingesters-within, the lookback calculation had to be
changed.

Previously, the lookback was calculated once at startup, since
query-ingesters-within was a global, unchanging value. With the move to
tenant overrides that can be changed at runtime, the lookback
calculation has been moved so it's done every time we get the ingesters
for querying.