Docs: clarify deprecation in "Parameter lifecycle"
We're moving towards a different definition of "deprecated."

### Current definition of "deprecated"

So far the definition that Mimir (and Cortex) has been using is to mean
that a feature or configuration parameter is effectively ignored.
Providing this parameter has no effect on Mimir. At the end of the
deprecation period the configuration parameter is removed and Mimir
fails to start if provided said parameter.

### New definition of "deprecated"

As per an internal design document the definition of "deprecated" should
 be more standardized across Grafana projects and should mean

> **Deprecation** occurs when a feature or product is planned for removal
> in a future release. In the period between deprecation and removal, it
> will be in maintenance.

### What this PR does

Updates docs to clarify what we mean by deprecated and what users can
expect from config parameters that are deprecated.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>